✨ feat: reformula página de atividades, layout responsivo e visão do supervisor
- Reformula página de atividades com modal de filtros multi-seleção
- Adiciona filtros por múltiplas categorias, projetos e módulos
- Adiciona layout responsivo com drawer mobile e navegação por acordeão
- Adiciona breakdown dev/sup/mtg na visão do supervisor
- Permite admin visualizar todos os setores no supervisor
- Adiciona auto-assignação de supervisor ao setor
- Atualiza rotas para todos os módulos

+++ app/Models/Activity.php
@@ -23,7 +23,7 @@ class Activity extends Model
             'ended_at' => 'datetime',
             'is_planned' => 'boolean',
             'energy_level' => 'integer',
-            'duration_minutes' => 'integer',
+            'duration_minutes' => 'float',
         ];
     }
